Interfaz Gráfica separada en varias clases
