+++ src/CMakeLists.txt
@@ -21,6 +21,8 @@ set(SRCS
 # TODO: Agregar aquí los archivos fuente de su proyecto
     main.cpp
     Tablero.cc
+    MainMenuDialog.cc
+    NewGameFrame.cc
 )
 
 if(APPLE)
